Allow compilation with clang (MLDB-1182)
This adds the toolchain={clang,gcc} option to the make system
(with gcc as the default) and fixes all compilation errors
and test failures under clang.  It required some modifications
to file layout as the clang linker is more picky about
resolved dependencies than gcc.  It significantly improves
the cleanliness and robustness of the code due to clang's
diagnistic messages.

There are some warnings that are disabled for clang which
could be re-enabled, but these tend to be low-impact (eg,
struct vs class for forward declatations).

It also uncovered one issue with the SVD library producing
invalid singular values, which was fixed.