Add 'finish' and logging script to mldb service
The docker init system (based on runsvdir) will call
/etc/service/mldb_runner/finish when mldb_runner exits.
The finish script will print a message with the signal number that killed the
process if it is available.

The 'run' script has been tweaked to 'exec' mldb_runner instead of forking it.
This is required for the finish script to properly get the signal code when
mldb_runner exits.

Since we now 'exec', we cannot use a pipeline to redirect stderr to stdout and
stdout to tee.  Instead, we use a logging script, which is a simple wrapper
around tee. runsv will arrange to wire up the file descriptor correctly
(mldb_runner's stdout to the logger's stdin).