tcmalloc not used by default (TRIVIAL)
In modern versions of Linux, tcmalloc's performance on 32+ threads is
much poorer than the built-in glibc malloc (ptmalloc2++).  Since MLDB
often runs on machines with 32+ virtual cores, it doesn't any longer
make sense to use tcmalloc and bring in an extra dependency.

Includes trivial fixes to tests that failed due to the change in
allocator.