Added direct use of indirect dependencies for static linking (OSX)
On Linux, we use linker flags to ensure that symbols found in one of
the dependencies of our dependent libraries are found without including
that library.  Not all linkers support this, so we list them explicitly
(this is particularly difficult with templates, as you need to ensure
that each symbol is instantiated in each of the libraries in which it
is used).