Added BitArray abstraction ready for JIT specialization
This BitArray abstraction creates a list of operations to implement each
operation to insert or extract elements from a BitArray, and implements
the operations by interpreting that list.  In the future, it will be
possible to JIT generate code to implement those operations for an ultra
efficient BitArray usage in a compute kernel.

Here we add the abstraction and tests; it will be used later on in
further commits.