32 bit fixed point accumulators
In most places, we store example weights in floats, which have a
23 bit mantissa.  Our fixed point accumulator has an effectvive 60
bit mantissa, which leads to most insignificant bits being zero
except for extremely large datasets.  This is the first step to
accumulating in 32 bits, which will save memory bandwidth and make
it more efficient to accumulate buckets where only 32 bit atomic
operations are available (as we don't need to synthesize 64 bit
atomic operations).