for_each_line rewrite for speed improvements; remove boost::iostreams
This is a rewrite of for_each_line to use a new implementation that is
able to better exploit the available parallelism, especially when
backed by content_descriptors with parallel-decompressible file formats
like lz4. There is much more comprehensive test support for the edge
cases, a modernization and simplification of the API and other
ergonomic improvements.

We also remove the dependency on boost::iostreams, including the
beginnings of support for including guard zones that allow the algorithms
to read a cache line or two beyond the end of the buffer without any
danger of segfaulting or security issues.

Notes on what else is included to support the headline items:

* A new compute_context class which allows a notion of priority in
  the tasks. This is currently barely exploited but will be important
  to tune performance in the future.
* A coalesced_range class that allows us to implement some algorithms
  over multiple memory buffers without needing to coalesce them.
* Support for some compressors to go back and overwrite the length
  field at the beginning of the file once the length is known. This
  allows the decompression to be more efficient when re-reading the
  file.